requirements.txt
streamlit
pandas
openpyxl
python-dateutil
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
requests
pydeck
PyPDF2
google-generativeai
google-cloud-aiplatform
plotly
anthropic
numpy
weasyprint
